# src/main/resources/openapi.yaml
openapi: "3.1.0"
info:
  title: "backend_flashpomo API"
  description: "backend_flashpomo API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
...
  /user/{userId}:
    put:
      summary: "PUT user/{userId}"
      operationId: "editUser"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserEditRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseUserEditDTO"
    delete:
      summary: "DELETE user/{userId}"
      operationId: "deleteUser"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseCommonDTO"
    get:
      summary: "GET user/{userId}"
      operationId: "getById"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GetByIdResponse"
...
components:
  schemas:
...
    UserEditRequestDTO:
      type: "object"
      properties:
        name:
          type: "string"
        email:
          type: "string"
        avatar:
          type: "array"
          items:
            type: "string"
            format: "byte"
        password:
          type: "string"
        role:
          type: "string"
        updatedAt:
          type: "string"
    ResponseUserEditDTO:
      type: "object"
      properties:
        name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        avatar:
          type: "array"
          items:
            type: "string"
            format: "byte"
        role:
          type: "string"
    ResponseCommonDTO:
      type: "object"
      properties:
        message:
          type: "string"
    GetByIdResponse:
      type: "object"
      properties:
        name:
          type: "string"
        email:
          type: "string"
...